==> 6_oof_features.ipynb <==
import os
import time 
import warnings
warnings.filterwarnings('ignore')

import nltk
import re
import pickle

%pylab inline
import matplotlib.pylab as plt
import seaborn as sns
plt.style.use('ggplot')
plt.style.use('seaborn-poster')
sns.set_palette('Set1', 10, desat=0.75)

import numpy as np
import pandas as pd
pd.set_option("display.max_rows", 200)
pd.set_option("display.max_colwidth", 200)

import scipy.sparse as sparse
from sklearn.decomposition import TruncatedSVD
from sklearn.preprocessing import MaxAbsScaler, MinMaxScaler



DATA_PATH = 'data/'

%%time
def load_sparse_csr(filename):
    "Loads scipy sparse matrix with csr format"
    loader = np.load(filename)
    return sparse.csr_matrix((loader['data'], loader['indices'], loader['indptr']),
                      shape=loader['shape'])

ytrain = pd.read_csv(os.path.join(DATA_PATH, 'target.csv')).target#[:10000]
ids = pd.read_csv(os.path.join(DATA_PATH, 'train_ids.csv'), usecols=['graph_id'])#[:10000]

data = load_sparse_csr(os.path.join(DATA_PATH, 'data_tfidf_stem_tags.npz'))#[:10000]
kagg = load_sparse_csr(os.path.join(DATA_PATH, 'kagg_tfidf_stem_tags.npz'))#[:10000]


from sklearn.metrics import log_loss
from sklearn.model_selection import KFold

import lightgbm as lgb
from fastFM.sgd import FMClassification

def lgb_pred(params, train, ytrain, valid, yvalid, test, kagg):
    train  = lgb.Dataset(train, ytrain)
    dvalid = lgb.Dataset(valid, yvalid, reference=train)
    
    gbm = lgb.train(params,train,
                    num_boost_round=100000,
                    valid_sets=[train, dvalid],
                    verbose_eval=False,
                    early_stopping_rounds=20)
    
    fold_pred  = gbm.predict(test,  num_iteration=gbm.best_iteration)
    kagg_pred  = gbm.predict(kagg,  num_iteration=gbm.best_iteration) 
    return fold_pred, kagg_pred

def fastfm_pred(params, train, ytrain, test, kagg):
    fmc = FMClassification(**params)
    fmc.fit(train, ytrain)
    fold_pred = fmc.predict_proba(test)
    kagg_pred = fmc.predict_proba(kagg)
    return fold_pred, kagg_pred
        
def get_oofs(model, data, kagg, y, ids, 
             n_splits=5, iters_total=3, params=None):
    """
    input:
        model: string ('lgb'/'fastfm'/'xgb'/'nnet') or any sklearn model
        data: train data in format of pandas DataFrame / numpay array / csr sparse matrix. All columns will be 
            used as features. Must not contain target or ids or non-numeric columns
        kagg: test data in format of pandas DataFrame / numpay array / csr sparse matrix. All columns will be 
            used as features. Must not contain target or ids or non-numeric columns
        y: target for train data in a form of pandas DataFrame / pandas Series / numpay array
        ids: graph ids in a form of pandas Series. Graphs ids are used to split train data into separate graphs 
            to prevent overfitting. In other words, quesitons of the same graph will always be in one fold.
        n_splits: number of splits for train data. Default value is 5. In order to get one OOF prediciton, 
            model must be fitted n_splits times
        iters_total: number of total iterations. Default value is 3. OOFs then will be blended. 
            Total number of times a model will be fitted is n_splits*iters_total.
        params: model parameters in a form of dictionary.
    output: 
        data_oofs: numpy array of OOF predictions for train data. Result is blended iters_total times
        kagg_oofs: numpy array of predictions for train data. Result is blended n_splits*iters_total times
    """
    
    if type(data).__name__=='DataFrame':
        data = data.values
    if type(kagg).__name__=='DataFrame':
        kagg = kagg.values
    if model == 'fastfm':
        y = y.replace(0, -1)
        if type(data).__name__ != 'csr_matrix':
            data = sparse.csr_matrix(data)
            kagg = sparse.csr_matrix(kagg)
    if type(y).__name__=='Series' or type(y).__name__=='DataFrame':
        y = y.values
    
    # matrices to store preditions
    data_oofs = np.zeros((data.shape[0]))
    kagg_oofs = np.zeros((kagg.shape[0]))
    
    graph_ids_unique = ids.graph_id.unique()
    
    for iter_num in range(iters_total):
        kf = KFold(n_splits=n_splits, shuffle=True)
        for train_graphs, test_graphs  in kf.split(graph_ids_unique):
            train_ind = ids[ids.graph_id.isin(graph_ids_unique[train_graphs])].index.values
            test_ind  = ids[ids.graph_id.isin(graph_ids_unique[test_graphs ])].index.values
            
            # Adding validation sets (from train set) for models, that require it. 
            # Validation size is 12.5% of train fold.
            if model=='xgb' or model=='lgb' or model=='nnet':
                kf_valid = KFold(n_splits=8, shuffle=True)
                graph_ids_train = graph_ids_unique[train_graphs]
                train_graphs, valid_graphs = list(kf_valid.split(graph_ids_train))[0]
                train_ind = ids[ids.graph_id.isin(graph_ids_train[train_graphs])].index.values
                valid_ind = ids[ids.graph_id.isin(graph_ids_train[valid_graphs])].index.values
                
            if model=='lgb':
                fold_pred, kagg_pred = lgb_pred (params, data[train_ind], y[train_ind], 
                                                         data[valid_ind], y[valid_ind],
                                                         data[test_ind], kagg)
            elif model=='xgb':
                fold_pred, kagg_pred = xgb_pred (params, data[train_ind], y[train_ind], 
                                                         data[valid_ind], y[valid_ind],
                                                         data[test_ind], kagg)
            elif model=='nnet':
                fold_pred, kagg_pred = nnet_pred(params, data[train_ind], y[train_ind], 
                                                         data[valid_ind], y[valid_ind],
                                                         data[test_ind], kagg)
            elif model=='fastfm':
                fold_pred, kagg_pred = fastfm_pred(params, data[train_ind], y[train_ind], data[test_ind], kagg)
               
            # Block for working with sklearn models
            else:
                model.fit(data[train_ind], y[train_ind])
                try:
                    fold_pred = model.predict_proba(data[test_ind])[:,1]
                    kagg_pred = model.predict_proba(kagg)[:,1]
                except:
                    try:
                        fold_pred = model.predict_proba(data[test_ind])
                        kagg_pred = model.predict_proba(kagg)
                    except:
                        fold_pred = model.predict(data[test_ind])
                        kagg_pred = model.predict(kagg)
            
            
            data_oofs[test_ind] += fold_pred
            kagg_oofs += kagg_pred
            print ('fold loss: ', log_loss(y[test_ind], fold_pred))            
        
        print ('iteration OOF score:', log_loss(y, data_oofs/(iter_num+1)))
    data_oofs /= iters_total
    kagg_oofs /= (iters_total*n_splits)
    return data_oofs, kagg_oofs

# OOF predicitons with LightGBM on TF-IDF transformed data.

%%time

# PCA transofmed matrices
pca = TruncatedSVD(n_components=20)
pca.fit(data)
data_pca = pca.transform(data)
kagg_pca = pca.transform(kagg)

# Empty dfs for oof preditions
data_preds = pd.DataFrame()
kagg_preds = pd.DataFrame()

# LightGBM hyperparameters
params = {
        'task': 'train','boosting_type': 'gbdt', 'objective': 'binary', 'metric': 'binary_logloss',
        'feature_fraction': 0.95,
        'min_data_in_leaf': 10, 
        'bagging_freq': 3, 
        'min_gain_to_split': 0, 
        'lambda_l2': 1, 
        'learning_rate': 0.075, 
        'num_leaves': 128, 
        'bagging_fraction': 0.85}

t_start = time.time()
data_preds['lgb_tfidf_oof'], kagg_preds['lgb_tfidf_oof'] = get_oofs('lgb'
                                                                    ,data
                                                                    ,kagg
                                                                    ,ytrain
                                                                    ,ids 
                                                                    ,params=params)
print ('lgb_tfidf_oof is done in {} minutes \n'.format(round((time.time()-t_start)/60,1) ))

t_start = time.time()
data_preds['lgb_tfidfpca_oof'], kagg_preds['lgb_tfidfpca_oof'] = get_oofs('lgb'
                                                                            ,data_pca
                                                                            ,kagg_pca
                                                                            ,ytrain
                                                                            ,ids
                                                                            ,params=params)
print ('lgb_tfidfpca_oof is done in {} minutes \n'.format(round((time.time()-t_start)/60,1) ))


data_preds.to_csv(os.path.join(DATA_PATH, 'train_lgb_tfidf_oof.csv'), index=False)
kagg_preds.to_csv(os.path.join(DATA_PATH, 'test_lgb_tfidf_oof.csv' ), index=False)

from keras.models import Sequential
from keras.layers import Dense, Dropout, Convolution1D, Activation
from keras.callbacks import EarlyStopping, ModelCheckpoint
from keras.optimizers import Adam, SGD, Adadelta
from keras.layers.advanced_activations import PReLU
from keras.utils.np_utils import to_categorical
from keras.regularizers import l2, l1
from keras.layers.normalization import BatchNormalization

def batch_generator(X, y, BATCH_SIZE, EPOCH_PARTION):
    """
    Batch generator for nnet training
    input:
        X - train dataset,  numpy array or csr matrix
        y - target, numpy array
        BATCH_SIZE - int, number of objects in batch. If X is csr matrix, it will be transformed 
        to dense array so batch size must be small enough for this array to fit in memory
        EPOCH_PARTION - float, share of objects that will be used for training in epoch
        
    Important: in this implementation each batch is constructed from random objects from train.
    """
    
    batch_number = 0
    batches_per_epoch = int(X.shape[0]/BATCH_SIZE*EPOCH_PARTION)
    while True:
        batch_indexes = np.random.choice(X.shape[0], BATCH_SIZE)
        if type(X).__name__ == 'csr_matrix':
            X_batch = X[batch_indexes].toarray()
        else:
            X_batch = X[batch_indexes]
        y_batch = to_categorical(y, num_classes=2)[batch_indexes]
        batch_number += 1
        yield X_batch, y_batch
        if batch_number == batches_per_epoch-1:
            batch_number = 0
            
def batch_generator_p(X, BATCH_SIZE):
    """
    Batch generator for nnet predicitons
    input:
        X - train dataset,  numpy array or csr matrix
        BATCH_SIZE - number of objects in batch. If X is csr matrix, it will be transformed 
        to dense array so batch size must be small enough for this array to fit in memory        
    """
    number_of_batches = X.shape[0] / np.ceil(X.shape[0]/BATCH_SIZE)
    batch_number = 0
    sample_index = np.arange(X.shape[0])
    while True:
        batch_indexes = sample_index[BATCH_SIZE*batch_number : BATCH_SIZE*(batch_number+1)]
        if type(X).__name__ == 'csr_matrix':
            X_batch = X[batch_indexes].toarray()
        else:
            X_batch = X[batch_indexes]
        batch_number += 1
        yield (X_batch)
        if batch_number == number_of_batches:
            batch_number = 0
            
def nnet(data, n1, n2, d1, d2, regul, **kwargs):
    """
    Function to compile simple nnet. Architecture is self-explanatory with code
    input:
        data - numpy arary or csr matrix for training
        n1, n2 - ints, number of neurons in first and second layers
        d1, d2 - float, dropouts in first and second layers
        regul - float, regularization paramter, the same for both layers
    output:
        nnet model
    """
    model = Sequential()
    if regul>0:
        model.add(Dense(n1, input_dim=data.shape[1], 
                        kernel_regularizer=l2(regul), 
                        activity_regularizer=l1(regul)))
    else:    
        model.add(Dense(n1, input_dim=data.shape[1]))
    model.add(PReLU())
    model.add(Dropout(d1))
    
    if regul>0:
        model.add(Dense(n2, 
                        kernel_regularizer=l2(regul), 
                        activity_regularizer=l1(regul)))
    else:    
        model.add(Dense(n2))
    model.add(PReLU())
    model.add(Dropout(d2))

    model.add(Dense(2))
    model.add(Activation('softmax'))
    
    adam = Adam()
    model.compile(loss='binary_crossentropy',
                  optimizer='adadelta',
                  metrics=['binary_crossentropy'])
    return model

def nnet_pred(params, train, ytrain, valid, yvalid, test, kagg):
    """
    
    """
    if type(train).__name__ == 'csr_matrix':
        scaler = MaxAbsScaler()
    else:
        scaler = MinMaxScaler()
    
    train = scaler.fit_transform(train)
    valid = scaler.transform(valid)
    test  = scaler.transform(test)
    kagg  = scaler.transform(kagg)
    
    model = nnet(train, **params)
    early_stopper = EarlyStopping(monitor='val_binary_crossentropy', patience=10, verbose=0, mode='auto')
    checkpoint = ModelCheckpoint(filepath='nnet_test.hdf5', monitor='val_binary_crossentropy', save_best_only=True)

    BATCH_SIZE = params.get('BATCH_SIZE', 256)
    EPOCH_PARTION = params.get('EPOCH_PARTION', 0.75)
    
    model.fit_generator(generator=batch_generator(train, ytrain, BATCH_SIZE, EPOCH_PARTION),
                        samples_per_epoch=int(train.shape[0]/BATCH_SIZE*EPOCH_PARTION),
                        verbose=0, nb_epoch=1000,

                        validation_data=batch_generator(valid, yvalid, BATCH_SIZE, EPOCH_PARTION), 
                        #nb_val_samples=int(valid.shape[0]/BATCH_SIZE*EPOCH_PARTION)*2, 
                        validation_steps = int(valid.shape[0]/BATCH_SIZE),

                        callbacks=[early_stopper, checkpoint])
    
    model.load_weights('nnet_test.hdf5') 
    model.compile(loss='binary_crossentropy', optimizer='adadelta', metrics=['binary_crossentropy'])
    
    fold_pred = model.predict_generator(generator=batch_generator_p(test, BATCH_SIZE), 
                                val_samples=test.shape[0]/BATCH_SIZE)
    kagg_pred = model.predict_generator(generator=batch_generator_p(kagg, BATCH_SIZE), 
                                val_samples=kagg.shape[0]/BATCH_SIZE)
    return fold_pred[:,1], kagg_pred[:,1]

data_preds = pd.DataFrame()
kagg_preds = pd.DataFrame()
params = {'n1':512, 'n2':256, 'd1':0.5, 'd2':0.25, 'regul':1e-9, 'BATCH_SIZE':512, 'EPOCH_PARTION':0.75}


t_start = time.time()
data_preds['nnet_tfidf_oof'], kagg_preds['nnet_tfidf_oof'] = get_oofs('nnet'
                                                            ,data
                                                            ,kagg
                                                            ,ytrain
                                                            ,ids 
                                                            ,params=params)
print ('nnet_tfidf_oof is done in {} minutes \n'.format(round((time.time()-t_start)/60,1) ))


data_preds.to_csv(os.path.join(DATA_PATH, 'train_nnet_tfidf_oof.csv'), index=False)
kagg_preds.to_csv(os.path.join(DATA_PATH, 'test_nnet_tfidf_oof.csv' ), index=False)

# OOF predicitons with factorization machines on TF-IDF transformed data.

%%time
data_preds = pd.DataFrame()
kagg_preds = pd.DataFrame()

params = {'n_iter':5000, 'init_stdev':5e-3, 'step_size':5e-3, 'l2_reg_w':10, 'l2_reg':10}
data_preds['FM_oof'], kagg_preds['FM_oof'] = get_oofs('fastfm', data, kagg, ytrain, ids, params=params)

data_preds.to_csv(os.path.join(DATA_PATH, 'train_fm_oof.csv'), index=False)
kagg_preds.to_csv(os.path.join(DATA_PATH, 'test_fm_oof.csv' ), index=False)

# END

https://www.kaggle.com/lystdo/lstm-with-word2vec-embeddings

import os

os.environ["KERAS_BACKEND"] = "tensorflow"

from keras.engine import Model
from keras.engine import Input

from sklearn.model_selection import StratifiedKFold
from sklearn.metrics import log_loss

from keras.layers import LSTM, concatenate
import numpy as np
from keras.callbacks import EarlyStopping, ModelCheckpoint
from keras.layers import Dense, Dropout, BatchNormalization
from keras.layers import Embedding
import pandas as pd

INPUT_FOLDER = '../../data'
TRAIN_MODE = True
NUM_EPOCHS = 10
NUM_FILTERS = 100
KERNEL_SIZE = 3
OUTPUT_SIZE = 128
MAX_FEATURES = 200000
EMBEDDING_SIZE = 300
DROPOUT_PROB = 0.5
MAX_LEN = 20
BATCH_SIZE = 64
PATIENCE = 3
WEIGHTS_NAME = 'weights_convnet.h5'
COMPILED_MODEL_NAME = 'compiled_convnet_model.h5'
TRAIN_OUTPUT_PROB_NAME = '../../data/train_convnet_proba.pkl'
GLOVE_VECTORDS_PATH = '../../data/glove.840B.300d.txt'
INPUT_TRAIN_FILE = '../../data/train.csv'
INPUT_TEST_FILE = '../../data/test.csv'
TEST_OUTPUT_PROB_NAME = '../../data/test_convnet_proba.pkl'
BASE_DIR = '../../data/'
EMBEDDING_FILE = BASE_DIR + 'GoogleNews-vectors-negative300.bin.gz'
TRAIN_DATA_FILE = BASE_DIR + 'train.csv'
TEST_DATA_FILE = BASE_DIR + 'test.csv'

# print('Processing text dataset')
#
#
# # The function "text_to_wordlist" is from
# # https://www.kaggle.com/currie32/quora-question-pairs/the-importance-of-cleaning-text
# def text_to_wordlist(text, remove_stopwords=False, stem_words=False):
#     # Clean the text, with the option to remove stopwords and to stem words.
#
#     # Convert words to lower case and split them
#     text = text.lower().split()
#
#     # Optionally, remove stop words
#     if remove_stopwords:
#         stops = set(stopwords.words("english"))
#         text = [w for w in text if not w in stops]
#
#     text = " ".join(text)
#
#     # Clean the text
#     text = re.sub(r"[^A-Za-z0-9^,!.\/'+-=]", " ", text)
#     text = re.sub(r"what's", "what is ", text)
#     text = re.sub(r"\'s", " ", text)
#     text = re.sub(r"\'ve", " have ", text)
#     text = re.sub(r"can't", "cannot ", text)
#     text = re.sub(r"n't", " not ", text)
#     text = re.sub(r"i'm", "i am ", text)
#     text = re.sub(r"\'re", " are ", text)
#     text = re.sub(r"\'d", " would ", text)
#     text = re.sub(r"\'ll", " will ", text)
#     text = re.sub(r",", " ", text)
#     text = re.sub(r"\.", " ", text)
#     text = re.sub(r"!", " ! ", text)
#     text = re.sub(r"\/", " ", text)
#     text = re.sub(r"\^", " ^ ", text)
#     text = re.sub(r"\+", " + ", text)
#     text = re.sub(r"\-", " - ", text)
#     text = re.sub(r"\=", " = ", text)
#     text = re.sub(r"'", " ", text)
#     text = re.sub(r"60k", " 60000 ", text)
#     text = re.sub(r":", " : ", text)
#     text = re.sub(r" e g ", " eg ", text)
#     text = re.sub(r" b g ", " bg ", text)
#     text = re.sub(r" u s ", " american ", text)
#     text = re.sub(r"\0s", "0", text)
#     text = re.sub(r" 9 11 ", "911", text)
#     text = re.sub(r"e - mail", "email", text)
#     text = re.sub(r"j k", "jk", text)
#     text = re.sub(r"\s{2,}", " ", text)
#
#     # Optionally, shorten words to their stems
#     if stem_words:
#         text = text.split()
#         stemmer = SnowballStemmer('english')
#         stemmed_words = [stemmer.stem(word) for word in text]
#         text = " ".join(stemmed_words)
#
#     # Return a list of words
#     return (text)
#
#
# texts_1 = []
# texts_2 = []
# YTrain = []
# with codecs.open(TRAIN_DATA_FILE) as f:
#     reader = csv.reader(f, delimiter=',')
#     header = next(reader)
#     for values in reader:
#         texts_1.append(text_to_wordlist(values[3]))
#         texts_2.append(text_to_wordlist(values[4]))
#         YTrain.append(int(values[5]))
# print('Found %s texts in train.csv' % len(texts_1))
#
# test_texts_1 = []
# test_texts_2 = []
# test_ids = []
# with codecs.open(TEST_DATA_FILE) as f:
#     reader = csv.reader(f, delimiter=',')
#     header = next(reader)
#     for values in reader:
#         test_texts_1.append(text_to_wordlist(values[1]))
#         test_texts_2.append(text_to_wordlist(values[2]))
#         test_ids.append(values[0])
# print('Found %s texts in test.csv' % len(test_texts_1))
#
# tokenizer = Tokenizer(num_words=MAX_FEATURES)
# tokenizer.fit_on_texts(texts_1 + texts_2 + test_texts_1 + test_texts_2)
#
# sequences_1 = tokenizer.texts_to_sequences(texts_1)
# sequences_2 = tokenizer.texts_to_sequences(texts_2)
# test_sequences_1 = tokenizer.texts_to_sequences(test_texts_1)
# test_sequences_2 = tokenizer.texts_to_sequences(test_texts_2)
#
# word_index = tokenizer.word_index
# print('Found %s unique tokens' % len(word_index))
#
# x1_train = pad_sequences(sequences_1, maxlen=MAX_LEN)
# x2_train = pad_sequences(sequences_2, maxlen=MAX_LEN)
# YTrain = np.array(YTrain)
# print('Shape of data tensor:', x1_train.shape)
# print('Shape of label tensor:', YTrain.shape)
#
# x1_test = pad_sequences(test_sequences_1, maxlen=MAX_LEN)
# x2_test = pad_sequences(test_sequences_2, maxlen=MAX_LEN)
# test_ids = np.array(test_ids)
#
# nb_words = min(MAX_FEATURES, len(word_index)) + 1
#
# with open('dump_sequences.pkl', 'wb') as f:
#     pickle.dump([x1_train, x2_train, x1_test, x2_test, nb_words], f)
#
# print('Preparing embedding matrix')
# word2vec = KeyedVectors.load_word2vec_format(EMBEDDING_FILE, binary=True)
# print('Found %s word vectors of word2vec' % len(word2vec.vocab))
# embedding_matrix = np.zeros((nb_words, EMBEDDING_SIZE))
# for word, i in word_index.items():
#     if word in word2vec.vocab:
#         embedding_matrix[i] = word2vec.word_vec(word)
# print('Null word embeddings: %d' % np.sum(np.sum(embedding_matrix, axis=1) == 0))
# np.save('embedding_matrix', embedding_matrix)

import cPickle as pickle
with open('dump_sequences.pkl', 'rb') as f:
    x1_train, x2_train, x1_test, x2_test, nb_words = pickle.load(f)
embedding_matrix = np.load('embedding_matrix.npy')
YTrain = pd.read_csv(INPUT_FOLDER + '/train.csv', usecols=['is_duplicate']).values.ravel()

class KerasConvnet:
    def __init__(self, use_weights=None):

        num_lstm = 267
        num_dense = 117
        rate_drop_lstm = 0.25
        rate_drop_dense = 0.15

        act = 'relu'

        embedding_layer = Embedding(nb_words,
                                    EMBEDDING_SIZE,
                                    weights=[embedding_matrix],
                                    input_length=MAX_LEN,
                                    trainable=False)
        lstm_layer = LSTM(num_lstm, dropout=rate_drop_lstm, recurrent_dropout=rate_drop_lstm)

        sequence_1_input = Input(shape=(MAX_LEN,), dtype='int32')
        embedded_sequences_1 = embedding_layer(sequence_1_input)
        x1 = lstm_layer(embedded_sequences_1)

        sequence_2_input = Input(shape=(MAX_LEN,), dtype='int32')
        embedded_sequences_2 = embedding_layer(sequence_2_input)
        y1 = lstm_layer(embedded_sequences_2)

        merged = concatenate([x1, y1])
        merged = Dropout(rate_drop_dense)(merged)
        merged = BatchNormalization()(merged)

        merged = Dense(num_dense, activation=act)(merged)
        merged = Dropout(rate_drop_dense)(merged)
        merged = BatchNormalization()(merged)
        preds = Dense(1, activation='sigmoid')(merged)

        model = Model(inputs=[sequence_1_input, sequence_2_input], outputs=preds)

        if use_weights is not None:
            model.load_weights(use_weights)
            
        model.compile(loss='binary_crossentropy',
                      optimizer='nadam',
                      metrics=['acc'])
        model.summary()

        self.model = model
        self.earlystop = EarlyStopping(monitor='val_loss', patience=3, verbose=0)
        self.checkpoint = ModelCheckpoint(WEIGHTS_NAME, monitor='val_loss', save_best_only=True, verbose=2)
        self.class_weight = {0: 1.309028344, 1: 0.472001959}

    def fit(self, train1, train2, y_train):
        print 'Fitting ... '
        input_arrays = [train1, train2]
        self.model.fit(input_arrays, y_train, batch_size=BATCH_SIZE,
                       epochs=NUM_EPOCHS, validation_split=0.1, shuffle=True,
                       callbacks=[self.earlystop, self.checkpoint])

    def predict(self, test1, test2):
        input_arrays = [test1, test2]
        preds = self.model.predict(input_arrays, batch_size=BATCH_SIZE, verbose=1)
        for _ in range(4):
            preds += self.model.predict(input_arrays, batch_size=BATCH_SIZE, verbose=1)
        preds /= 5.

        return preds

print x1_train.shape, x2_train.shape
print x1_test.shape, x2_test.shape
print nb_words

N_FOLDS = 3
skf = StratifiedKFold(n_splits=N_FOLDS).split(np.zeros((x1_train.shape[0], 1)), YTrain)
splits = list(skf)

print "Creating train and test sets for blending."

preds_train = np.zeros((x1_train.shape[0],))
loglosses = []

preds_test_j = np.zeros((x1_test.shape[0], N_FOLDS))
for fold_id, (train_indexes, predict_indexes) in enumerate(splits):
    print "Fold", fold_id

    clf = KerasConvnet()
    clf.fit(x1_train[train_indexes], x2_train[train_indexes], YTrain[train_indexes])

    y_pred = clf.predict(x1_train[predict_indexes], x2_train[predict_indexes])
    preds_train[predict_indexes] = y_pred

    lloss = log_loss(YTrain[predict_indexes], y_pred)
    loglosses.append(lloss)
    print 'LogLoss: ', lloss

    # Predict on entire test set
    preds_test_j[:, fold_id] = clf.predict(x1_test, x2_test).ravel()

print "Out of fold logloss-es:\n", loglosses

# Save OOF predictions as features
pd.DataFrame(preds_train, columns=['siamese_oof']).to_csv(INPUT_FOLDER + '/features/siamese_oof_train.csv')

# Save mean of predictions for test
preds_test = preds_test_j.mean(1)
pd.DataFrame(preds_test, columns=['siamese_oof']).to_csv(INPUT_FOLDER + '/features/siamese_oof_test.csv')
